spearman

+++ src/task+text_emb/run_taskemb_QA.py
@@ -449,7 +449,7 @@ def main():
     model = model_class.from_pretrained(args.model_name_or_path,
                                         from_tf=bool('.ckpt' in args.model_name_or_path),
                                         config=config,
-                                        cache_dir=args.cache_dir if args.cache_dir else None, torch_dtype=torch.float16)
+                                        cache_dir=args.cache_dir if args.cache_dir else None)
 
     model.to(args.device)
